+++ alidns-20150109/alibabacloud_alidns20150109/client.py
@@ -8232,6 +8232,8 @@ class Client(OpenApiClient):
         query = {}
         if not UtilClient.is_unset(request.app_key_id):
             query['AppKeyId'] = request.app_key_id
+        if not UtilClient.is_unset(request.auth_code):
+            query['AuthCode'] = request.auth_code
         if not UtilClient.is_unset(request.lang):
             query['Lang'] = request.lang
         req = open_api_models.OpenApiRequest(
@@ -8262,6 +8264,8 @@ class Client(OpenApiClient):
         query = {}
         if not UtilClient.is_unset(request.app_key_id):
             query['AppKeyId'] = request.app_key_id
+        if not UtilClient.is_unset(request.auth_code):
+            query['AuthCode'] = request.auth_code
         if not UtilClient.is_unset(request.lang):
             query['Lang'] = request.lang
         req = open_api_models.OpenApiRequest(